requirements-azure.txt
azure-sdk-for-python>=1.13.0
azure-identity>=1.14.0
azure-ai-language-conversations>=1.1.0
azure-ai-textanalytics>=5.3.0
openai>=1.3.0
python-dotenv>=1.0.0
